Move OrderedSerialization into zero-dep scalding-serialization module
