+++ scalding-serialization/src/main/scala/com/twitter/scalding/serialization/JavaStreamEnrichments.scala
