+++ scalding-serialization/src/main/scala/com/twitter/scalding/serialization/Reader.scala
