+++ scalding-serialization/src/main/scala/com/twitter/scalding/serialization/StringOrderedSerialization.scala
