+++ scalding-serialization/src/test/scala/com/twitter/scalding/serialization/SerializationProperties.scala
