+++ scalding-serialization/src/test/scala/com/twitter/scalding/serialization/WriterReaderProperties.scala
